refactor(bpjson): split the NodesForm
